requirements.txt
gspread
google-auth
slack-sdk
python-dateutil